added dividends

+++ src/models/moex_models/dividend.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 from pydantic import BaseModel
 from typing import Optional
 from datetime import date
@@ -7,7 +6,7 @@ from datetime import date
 class Dividend(BaseModel):
     secid: str
     record_date: date
-    value: Decimal
+    value: float
 
     action_type: str = "dividend"
